Make CCC charts not wait for svgweb when unnecessary
CCC chart components were using pv.listForPageLoad
to wait for all scripts to be loaded (so svgweb was
guaranteed to be available), and this was causing
the chart execution to happen asynchronously with
the dashboard. This seriously screwed up the way
Mobile resized the charts, so we made the chart
only wait for svgweb if native svg is not available.