Update the resource-fetching logic
We had some issues with the security checks in the
resource handling code. One was that we allowed
fetching files from the system folder. Now we only
allow files from /system/ if they fall within our
plugin.

Another was that the allowed roots were a bit
inconsistent, which was creating some issues in
dashboards using getResource instead of res.